feat: remove unused speaker setup in test app
